absorbing userForm into profile page
