--- client/src/app/forms/userform/userform.component.html
@@ -1,40 +1,0 @@
-<div class="formContainer" fxLayoutAlign="center center" flxLayout="flex-container" fxLayout="column">
-
-    <form [formGroup]="userProfileForm" (ngSubmit)="onProfileUpdate(userProfileForm.value)">
-
-        <!-- username entry -->
-        <mat-form-field appearance="fill" class="flex-item">
-            <mat-label>Username</mat-label>
-            <input matNativeControl placeholder="username" formControlName="username" matInput>
-        </mat-form-field>
-        <br>
-        <!-- email entry - autofilled from login? -->
-        <mat-form-field appearance="fill" class="flex-item">
-            <mat-label>Email</mat-label>
-            <input matNativeControl placeholder="email" formControlName="email" matInput>
-            <mat-error *ngIf="email.invalid">{{getErrorMessage()}}</mat-error>
-        </mat-form-field>
-        <br>
-        <!-- preferred pronoun entry with dropdown menu -->
-        <mat-form-field appearance="fill" class="flex-item">
-            <mat-label>Preferred Pronoun</mat-label>
-            <!-- <mat-select formControlName="pronouns" [(value)]="pronounSelection">
-                <mat-option placeholder="Select One" value="she/her/hers">she/her/hers</mat-option>
-                <mat-option placeholder="Select One" value="he/him/his">he/him/his</mat-option>
-                <mat-option placeholder="Select One" value="they/them/theirs">they/them/theirs</mat-option>
-            </mat-select> -->
-            <input matNativeControl placeholder="pronouns" formControlName="pronouns" matInput>
-        </mat-form-field>
-        <br>
-        <!-- location entry -->
-        <mat-form-field appearance="fill" class="flex-item">
-            <mat-label>Location</mat-label>
-            <input matNativeControl placeholder="location" formControlName="location" matInput>
-        </mat-form-field>
-
-        <div fxLayout="row" fxLayoutAlign="center center">
-            <!-- submit button -->
-            <button mat-raised-button type="submit" [disabled]="!userProfileForm.valid">Submit</button>
-        </div>
-    </form>
-</div>
