--- client/src/app/forms/userform/userform.component.scss
@@ -1,10 +1,0 @@
-.formContainer {
-    min-height: 90vh;
-    max-width: 800px;
-    padding-bottom: 50px;
-}
-
-.flex-item {
-    margin: 10px;
-    width: 50vw;
-}
