+++ client/src/app/pages/profile/profile.component.html
@@ -3,11 +3,42 @@
     fxLayout.xs="row wrap">
 
     <mat-card fxLayout="column" fxLayoutAlign="space-around center">
-        <mat-card-title *ngIf="user" class="text">{{ user.name }}</mat-card-title>
-        <img mat-card-image *ngIf="user" [src]="user.picture" alt="Profile Picture">
+        <mat-card-title *ngIf="user" class="text">{{ user.username}}</mat-card-title>
+        <img mat-card-image *ngIf="user" [src]="user.photo" alt="Profile Picture">
         <mat-card-subtitle *ngIf="user" class="text">{{ user.email }}</mat-card-subtitle>
     </mat-card>
-
-    <app-userform></app-userform>
-
+    
+    <form [formGroup]="userProfileForm" (ngSubmit)="onProfileUpdate(userProfileForm.value)" fxLayout="column"
+        fxLayoutAlign="center center" flxLayout="flex-container">
+    
+        <!-- username entry -->
+        <mat-form-field appearance="fill" class="flex-item">
+            <mat-label>Username</mat-label>
+            <input matNativeControl placeholder="username" formControlName="username" matInput>
+        </mat-form-field>
+    
+        <!-- email entry - autofilled from login? -->
+        <mat-form-field appearance="fill" class="flex-item">
+            <mat-label>Email</mat-label>
+            <input matNativeControl placeholder="email" formControlName="email" matInput>
+            <mat-error *ngIf="email.invalid">{{getErrorMessage()}}</mat-error>
+        </mat-form-field>
+    
+        <!-- preferred pronoun entry with dropdown menu -->
+        <mat-form-field appearance="fill" class="flex-item">
+            <mat-label>Preferred Pronoun</mat-label>
+            <input matNativeControl placeholder="pronouns" formControlName="pronouns" matInput>
+        </mat-form-field>
+    
+        <!-- location entry -->
+        <mat-form-field appearance="fill" class="flex-item">
+            <mat-label>Location</mat-label>
+            <input matNativeControl placeholder="location" formControlName="location" matInput>
+        </mat-form-field>
+    
+        <div fxLayout="row" fxLayoutAlign="center center">
+            <!-- submit button -->
+            <button mat-raised-button type="submit" [disabled]="!userProfileForm.valid">Submit</button>
+        </div>
+    </form>
 </div>
